provide machine ids to SchedulingEventListeners
it is not uncommon that VM ids are unique in a mesos cluster, and in
many cases are better candidates to index machines than hostnames, which
are usually IPs that can be reused in the Cloud.

Pass down both hostnames and VM ids in SchedulingResults, so callbacks
can choose which one is a better fit for their usage.